requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.3
pandas==0.23.4
psycopg2-binary==2.7.5
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.0
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
